Create a function exp. equal to the variable numba
Give the function a parameter of 'string'.
When a string shorter than 7 characters is passed to the function, print
"What a short little word." When the string is exactly 7 characters,
print, "7, what a perfect choice!" When the string is more than 7
characters, print, "I'm sorry, but that's too many to count."
Then pass parameters of '7', 'seventy', and 'university' consecutively.